[level 0] Title: 두 수의 나눗셈, Time: 0.01 ms, Memory: 4.2 MB -BaekjoonHub

+++ 프로그래머스/0/120806. 두 수의 나눗셈/두 수의 나눗셈.c
@@ -3,7 +3,7 @@
 #include <stdlib.h>
 
 int solution(int num1, int num2) {
-    float answer = 0;
+    int answer = 0;
     answer = ((float)num1 / num2) * 1000;
     return (int)answer;
 }
